Separando o código assembly
Separei o código gerar_assembly em vários subcódigos dentro da pasta raiz/src/assembly. O objetivo foi simplificar o código assembly em diferentes partes e juntando tudo com o código builder.py
