+++ outputs/assembly/op_1.S
@@ -1,6 +1,6 @@
 ; ====================================================================
 ; Código Assembly gerado automaticamente para Arduino Uno (ATmega328p)
-; Processador de expressões RPN (Reverse Polish Notation) - 16-BIT VERSION
+; Processador RPN - 16-BIT VERSION
 ; Suporte para inteiros de 0 a 65535
 ; Compilado com PlatformIO/AVR-GCC
 ; ====================================================================
@@ -14,13 +14,9 @@
 ; SEÇÃO DE DADOS - 16-BIT VERSION
 ; ====================================================================
 
-; Stack pointer para pilha de inteiros 16-bit (simula pilha RPN)
-; Cada entrada da pilha ocupa 2 bytes (16 bits)
-; Memória SRAM para armazenar resultados e variáveis
-
 .section .data
 stack_ptr: .byte 1        ; Ponteiro da pilha RPN
-mem_vars:  .space 52      ; Espaço para 26 variáveis de 16-bit (A-Z)
+mem_vars:  .space 52      ; 26 variáveis 16-bit (A-Z)
 temp_result: .space 4     ; Resultado temporário
 
 .section .text
@@ -36,10 +32,8 @@ main:
     ldi r16, 0x08
     out SPH, r16
 
-    ; Inicializar UART para debug (9600 baud)
+    ; Inicializar UART (9600 baud) e pilha RPN
     rcall uart_init
-
-    ; Inicializar pilha RPN
     rcall stack_init
 
     ; Processar expressão RPN
@@ -57,7 +51,7 @@ main:
 
 processar_rpn:
     ; Processando expressão RPN com suporte 16-bit:
-    ; Expressão: 6 2 *
+    ; Expressão: 4 2 +
 
     ; Debug: Enviar mensagem inicial
     ldi r16, 'S'
@@ -75,18 +69,18 @@ processar_rpn:
     ldi r16, 10
     rcall uart_transmit
 
-    ; Processando token 0: '6'
-    ; Debug: Enviar "P6"
+    ; Processando token 0: '4'
+    ; Debug: Enviar "P4"
     ldi r16, 'P'
     rcall uart_transmit
-    ldi r16, '6'
+    ldi r16, '4'
     rcall uart_transmit
     ldi r16, ' '
     rcall uart_transmit
 
-    ldi r16, 6      ; Byte baixo do valor 6
-    ldi r17, 0 ; Byte alto do valor 6
-    rcall stack_push_int         ; Empilha valor inteiro 16-bit
+    ldi r16, 4      ; Byte baixo
+    ldi r17, 0 ; Byte alto
+    rcall stack_push_int
 
     ; Processando token 1: '2'
     ; Debug: Enviar "P2"
@@ -97,50 +91,121 @@ processar_rpn:
     ldi r16, ' '
     rcall uart_transmit
 
-    ldi r16, 2      ; Byte baixo do valor 2
-    ldi r17, 0 ; Byte alto do valor 2
-    rcall stack_push_int         ; Empilha valor inteiro 16-bit
+    ldi r16, 2      ; Byte baixo
+    ldi r17, 0 ; Byte alto
+    rcall stack_push_int
 
-    ; Processando token 2: '*'
-    ; Debug: Enviar "MUL"
-    ldi r16, 'M'
+    ; Processando token 2: '+'
+    ; Debug: Enviar "ADD"
+    ldi r16, 'A'
     rcall uart_transmit
-    ldi r16, 'U'
+    ldi r16, 'D'
     rcall uart_transmit
-    ldi r16, 'L'
+    ldi r16, 'D'
     rcall uart_transmit
     ldi r16, ' '
     rcall uart_transmit
 
     rcall stack_pop_int      ; Remove segundo operando
-    push r16                 ; Salvar r16 antes do debug
-    push r17                 ; Salvar r17 antes do debug
+
+    ; Debug do segundo operando
+    push r16
+    push r17
+    ldi r16, 'B'
+    rcall uart_transmit
+    ldi r16, ':'
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
+    push r17
+    mov r16, r17
+    rcall send_byte_as_hex
+    pop r17
+    pop r16
+    push r16
+    push r17
+    rcall send_byte_as_hex
+    ldi r16, ' '
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
+    push r17
     rcall send_number_16bit
     ldi r16, ' '
     rcall uart_transmit
-    pop r17                  ; Restaurar r17
-    pop r16                  ; Restaurar r16
-    mov r18, r16             ; Move para r18:r19
+    pop r17
+    pop r16
+
+    mov r18, r16             ; guarda segundo operando
     mov r19, r17
+
     rcall stack_pop_int      ; Remove primeiro operando
-    push r16                 ; Salvar r16 antes do debug
-    push r17                 ; Salvar r17 antes do debug
+
+    ; Debug do primeiro operando
+    push r16
+    push r17
+    ldi r16, 'A'
+    rcall uart_transmit
+    ldi r16, ':'
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
+    push r17
+    mov r16, r17
+    rcall send_byte_as_hex
+    pop r17
+    pop r16
+    push r16
+    push r17
+    rcall send_byte_as_hex
+    ldi r16, ' '
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
+    push r17
     rcall send_number_16bit
     ldi r16, ' '
     rcall uart_transmit
-    pop r17                  ; Restaurar r17
-    pop r16                  ; Restaurar r16
-    rcall multiply_int       ; TRUE 16-bit multiplication
-    push r16                 ; Salvar resultado
+    pop r17
+    pop r16
+
+    ; Soma 16-bit
+    add r16, r18
+    adc r17, r19
+
+    ; Debug do resultado
+    push r16
+    push r17
+    ldi r16, 'R'
+    rcall uart_transmit
+    ldi r16, ':'
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
+    push r17
+    rcall send_byte_as_hex
+    mov r16, r17
+    rcall send_byte_as_hex
+    ldi r16, ' '
+    rcall uart_transmit
+    pop r17
+    pop r16
+    push r16
     push r17
     rcall send_number_16bit
     ldi r16, 13
     rcall uart_transmit
     ldi r16, 10
     rcall uart_transmit
-    pop r17                  ; Restaurar resultado
+    pop r17
     pop r16
-    rcall stack_push_int     ; Empilha resultado
+
+    rcall stack_push_int
 
     ; Fim do processamento
     ret
