+++ outputs/assembly/op_9.S
@@ -1,6 +1,6 @@
 ; ====================================================================
 ; Código Assembly gerado automaticamente para Arduino Uno (ATmega328p)
-; Processador de expressões RPN (Reverse Polish Notation) - 16-BIT VERSION
+; Processador RPN - 16-BIT VERSION
 ; Suporte para inteiros de 0 a 65535
 ; Compilado com PlatformIO/AVR-GCC
 ; ====================================================================
@@ -14,13 +14,9 @@
 ; SEÇÃO DE DADOS - 16-BIT VERSION
 ; ====================================================================
 
-; Stack pointer para pilha de inteiros 16-bit (simula pilha RPN)
-; Cada entrada da pilha ocupa 2 bytes (16 bits)
-; Memória SRAM para armazenar resultados e variáveis
-
 .section .data
 stack_ptr: .byte 1        ; Ponteiro da pilha RPN
-mem_vars:  .space 52      ; Espaço para 26 variáveis de 16-bit (A-Z)
+mem_vars:  .space 52      ; 26 variáveis 16-bit (A-Z)
 temp_result: .space 4     ; Resultado temporário
 
 .section .text
@@ -36,10 +32,8 @@ main:
     ldi r16, 0x08
     out SPH, r16
 
-    ; Inicializar UART para debug (9600 baud)
+    ; Inicializar UART (9600 baud) e pilha RPN
     rcall uart_init
-
-    ; Inicializar pilha RPN
     rcall stack_init
 
     ; Processar expressão RPN
@@ -57,7 +51,7 @@ main:
 
 processar_rpn:
     ; Processando expressão RPN com suporte 16-bit:
-    ; Expressão: 2 RES
+    ; Expressão: 1 RES
 
     ; Debug: Enviar mensagem inicial
     ldi r16, 'S'
@@ -75,21 +69,21 @@ processar_rpn:
     ldi r16, 10
     rcall uart_transmit
 
-    ; Processando token 0: '2'
-    ; Debug: Enviar "P2"
+    ; Processando token 0: '1'
+    ; Debug: Enviar "P1"
     ldi r16, 'P'
     rcall uart_transmit
-    ldi r16, '2'
+    ldi r16, '1'
     rcall uart_transmit
     ldi r16, ' '
     rcall uart_transmit
 
-    ldi r16, 2      ; Byte baixo do valor 2
-    ldi r17, 0 ; Byte alto do valor 2
-    rcall stack_push_int         ; Empilha valor inteiro 16-bit
+    ldi r16, 1      ; Byte baixo
+    ldi r17, 0 ; Byte alto
+    rcall stack_push_int
 
     ; Processando token 1: 'RES'
-    rcall comando_res     ; Comando RES
+    rcall comando_res
 
     ; Fim do processamento
     ret
