+++ outputs/tokens/tokens_gerados.txt
@@ -1,10 +1,10 @@
-6 2 *
-12 4 /
-9 5 -
-3 3 ^
-15 4 %
-2 2 + 5 1 - *
-7 MEM
+4 2 +
+6 3 /
+5 2 ^
+7 2 %
+8 3 -
+2 3 * 4 5 + /
+9 MEM
 MEM
-2 RES
-3 2 + 10 5 / *
+1 RES
+1 2 + 3 4 + *
