+++ src/functions/assembly/__init__.py
@@ -1,0 +1,5 @@
+from .builder import gerarAssembly
+from .io import save_assembly
+from .registers import save_registers_inc
+
+__all__ = ["gerarAssembly", "save_assembly", "save_registers_inc"]
