+++ src/functions/assembly/operations.py
@@ -1,0 +1,392 @@
+# operations.py
+from typing import List
+
+# -----------------------------
+# Helpers de análise de tokens
+# -----------------------------
+
+def is_number(token: str) -> bool:
+    try:
+        float(token)
+        return True
+    except ValueError:
+        return False
+
+def is_integer(token: str) -> bool:
+    try:
+        val = float(token)
+        return val == int(val)
+    except ValueError:
+        return False
+
+def is_variable_mem(token: str) -> bool:
+    return token.isalpha() and token.isupper() and token not in ("MEM", "RES")
+
+# ---------------------------------
+# Geração de PUSH (público + interno)
+# ---------------------------------
+
+def _gerar_push_int_com_debug(valor: int) -> List[str]:
+    digit_char = str(valor) if valor < 10 else 'X'
+    return [
+        f"    ; Debug: Enviar \"P{valor}\"",
+        "    ldi r16, 'P'",
+        "    rcall uart_transmit",
+        f"    ldi r16, '{digit_char}'",
+        "    rcall uart_transmit",
+        "    ldi r16, ' '",
+        "    rcall uart_transmit",
+        "",
+        f"    ldi r16, {valor & 0xFF}      ; Byte baixo",
+        f"    ldi r17, {(valor >> 8) & 0xFF} ; Byte alto",
+        "    rcall stack_push_int",
+        ""
+    ]
+
+def gerar_push_int(valor: int) -> List[str]:
+    """Wrapper público (não recursivo)."""
+    return _gerar_push_int_com_debug(valor)
+
+# ---------------------------------
+# Geração de operações aritméticas
+# ---------------------------------
+
+def _operacao_map() -> dict[str, List[str]]:
+    return {
+        '+': [
+            "    ; Debug: Enviar \"ADD\"",
+            "    ldi r16, 'A'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'D'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'D'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int      ; Remove segundo operando",
+            "",
+            "    ; Debug do segundo operando",
+            "    push r16",
+            "    push r17",
+            "    ldi r16, 'B'",
+            "    rcall uart_transmit",
+            "    ldi r16, ':'",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    mov r16, r17",
+            "    rcall send_byte_as_hex",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_byte_as_hex",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "",
+            "    mov r18, r16             ; guarda segundo operando",
+            "    mov r19, r17",
+            "",
+            "    rcall stack_pop_int      ; Remove primeiro operando",
+            "",
+            "    ; Debug do primeiro operando",
+            "    push r16",
+            "    push r17",
+            "    ldi r16, 'A'",
+            "    rcall uart_transmit",
+            "    ldi r16, ':'",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    mov r16, r17",
+            "    rcall send_byte_as_hex",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_byte_as_hex",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "",
+            "    ; Soma 16-bit",
+            "    add r16, r18",
+            "    adc r17, r19",
+            "",
+            "    ; Debug do resultado",
+            "    push r16",
+            "    push r17",
+            "    ldi r16, 'R'",
+            "    rcall uart_transmit",
+            "    ldi r16, ':'",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_byte_as_hex",
+            "    mov r16, r17",
+            "    rcall send_byte_as_hex",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "",
+            "    rcall stack_push_int",
+            ""
+        ],
+        '-': [
+            "    ; Debug: Enviar \"SUB\"",
+            "    ldi r16, 'S'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'U'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'B'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int      ; b",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    mov r18, r16",
+            "    mov r19, r17",
+            "    rcall stack_pop_int      ; a",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "",
+            "    sub r16, r18",
+            "    sbc r17, r19",
+            "",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall stack_push_int",
+            ""
+        ],
+        '*': [
+            "    ; Debug: Enviar \"MUL\"",
+            "    ldi r16, 'M'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'U'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'L'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    mov r18, r16",
+            "    mov r19, r17",
+            "    rcall stack_pop_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall multiply_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall stack_push_int",
+            ""
+        ],
+        '/': [
+            "    ; Debug: Enviar \"DIV\"",
+            "    ldi r16, 'D'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'I'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'V'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int      ; divisor",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    mov r18, r16",
+            "    mov r19, r17",
+            "    rcall stack_pop_int      ; dividendo",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall divide_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall stack_push_int",
+            ""
+        ],
+        '%': [
+            "    ; Debug: Enviar \"MOD\"",
+            "    ldi r16, 'M'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'O'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'D'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int      ; divisor",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    mov r18, r16",
+            "    mov r19, r17",
+            "    rcall stack_pop_int      ; dividendo",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall modulo_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall stack_push_int",
+            ""
+        ],
+        '^': [
+            "    ; Debug: Enviar \"POW\"",
+            "    ldi r16, 'P'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'O'",
+            "    rcall uart_transmit",
+            "    ldi r16, 'W'",
+            "    rcall uart_transmit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "",
+            "    rcall stack_pop_int      ; expoente",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    mov r18, r16",
+            "    mov r19, r17",
+            "    rcall stack_pop_int      ; base",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, ' '",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall power_int",
+            "    push r16",
+            "    push r17",
+            "    rcall send_number_16bit",
+            "    ldi r16, 13",
+            "    rcall uart_transmit",
+            "    ldi r16, 10",
+            "    rcall uart_transmit",
+            "    pop r17",
+            "    pop r16",
+            "    rcall stack_push_int",
+            ""
+        ],
+    }
+
+def gerar_operacao(operador: str) -> List[str]:
+    """Retorna as linhas Assembly para o operador informado."""
+    mapa = _operacao_map()
+    return mapa.get(operador, [f"    ; Operação {operador} não implementada", ""])
+
+__all__ = [
+    "is_number", "is_integer", "is_variable_mem",
+    "gerar_push_int", "gerar_operacao",
+]
